models/pyproject.toml
[project]
name = "aesfa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "av>=14.3.0",
    "coremltools>=8.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "onnx>=1.17.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "seaborn>=0.13.2",
    "thop>=0.1.1.post2209072238",
    "torch==2.5.0",
    "torchvision>=0.20.0",
]
